+++ src/constants.tsx
@@ -1,7 +1,3 @@
 export const ADDON_ID = 'storybook/paddings';
 export const PARAM_KEY = 'paddings';
 export const DEFAULT_PADDING = '';
-
-export const EVENTS = {
-  UPDATE: `${ADDON_ID}/update`,
-};
