+++ src/helpers.tsx
@@ -1,19 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 import { logger } from '@storybook/client-logger';
-import { WrapperSettings } from '@storybook/addons';
 import { DEFAULT_PADDING } from './constants';
-
-type Padding = { name: string; value: string };
-
-export type PaddingsParameter = {
-  default?: boolean;
-  disable?: boolean;
-  values: Padding[];
-};
-
-export type PaddingWithDefault = Padding & Pick<PaddingsParameter, 'default'>;
-
-type Options = Padding[] | PaddingsParameter | WrapperSettings['parameters'];
+import { PaddingWithDefault, Options, Padding } from './types';
 
 export const isEnabled = (values: PaddingWithDefault[]): boolean =>
   values.length > 0;
@@ -21,7 +9,7 @@ export const isEnabled = (values: PaddingWithDefault[]): boolean =>
 export const normalizeValues = (options: Options): PaddingWithDefault[] => {
   if (Array.isArray(options)) {
     logger.warn(
-      'Using array parameter with Paddings addon is deprecated, and support will be dropped in a future major version. Please refer to the release notes for migration: https://github.com/rbardini/storybook-addon-paddings/releases/tag/v2.1.0',
+      'Using array parameter with Paddings addon is deprecated, and support will be dropped in a future major version. Please refer to the release notes for migration: https://github.com/rbardini/storybook-addon-paddings/releases/tag/v3.1.0',
     );
     return options;
   }
